Add support for adding attachment for transactions
Feature request #32

+++ app/src/main/java/xyz/hisname/fireflyiii/repository/models/transaction/TransactionAttributes.kt
@@ -14,7 +14,7 @@ data class TransactionAttributes(
         @SerializedName("type")
         val transactionType: String,
         val identifier: Int,
-        val journal_id: Int,
+        val journal_id: Long,
         val journal_description: String,
         val reconciled: Boolean,
         val amount: Double,
